Separated Storage Matrix graph implementation into multiple files.
